% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incompleteExample}
\alias{incompleteExample}
\title{Example Partial or Incomplete Surface}
\format{
A data frame with 64 rows and 4 columns
\describe{
\item{concA}{The concentration of drug A}
\item{concB}{The concentration of drug B}
\item{truth}{The true response surface value at the given dose pair}
\item{measure}{The sampled noisy measurement of the response surface at
the given dose pair}
}
}
\usage{
incompleteExample
}
\description{
A synthetically generated response surface using parameter vector describing
a surface with one only barely detectable effect. The surface was generated
with IDMA of 1, IDMB of 100, na of 3, nb of 3, kappa of 0, E0 of 0, EfA of 1,
EfB of 0.1, and Ef of 1. Every pair of concentrations is sampled once, with
concentrations of 0 and a seven-point two-fold dilution from 0.125 to 8.
"Measurements" were sampled from a normal noise distribution around ground
truth values with a standard deviation of 0.07.
}
\keyword{datasets}
